requirements.txt
mcp[cli]>=1.2.0
httpx>=0.27.0
python-dotenv>=1.0.0
openai>=1.0.0
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.0.0